// src/domain/enums/notification-type.enum.ts
export enum NotificationType {
  APPOINTMENT_CONFIRMED = 'APPOINTMENT_CONFIRMED',
  APPOINTMENT_REMINDER = 'APPOINTMENT_REMINDER',
  APPOINTMENT_CANCELLED = 'APPOINTMENT_CANCELLED',
  APPOINTMENT_RESCHEDULED = 'APPOINTMENT_RESCHEDULED',
  PAYMENT_RECEIVED = 'PAYMENT_RECEIVED',
  PAYMENT_FAILED = 'PAYMENT_FAILED',
  PROMO_CODE = 'PROMO_CODE',
  WELCOME = 'WELCOME',
  PASSWORD_RESET = 'PASSWORD_RESET',
  SYSTEM = 'SYSTEM',
  FOLLOW_UP = 'FOLLOW_UP',
  REACTIVATION = 'REACTIVATION',
  LOYALTY_REMINDER = 'LOYALTY_REMINDER',
  MEMBERSHIP_RENEWAL = 'MEMBERSHIP_RENEWAL',
  RECURRING_TREATMENT = 'RECURRING_TREATMENT',
}